requirements.txt
pybaseball
duckdb
pandas
dbt-duckdb
jupyter
requests
python-dotenv
